# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP metadata capture service with synchronous and asynchronous inventory flows.",
        "title": "Metadata Inventory API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Returns process liveness for container orchestration and smoke tests.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Liveness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HealthResponse"
                        }
                    }
                }
            }
        },
        "/ready": {
            "get": {
                "description": "Verifies that MongoDB and Kafka are reachable by the API service.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Readiness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReadyResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReadyResponse"
                        }
                    }
                }
            }
        },
        "/v1/metadata": {
            "get": {
                "description": "Returns cached metadata when present. On cache miss, persists a pending record, dispatches async collection, and returns 202 immediately.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Get metadata for a URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Absolute URL to look up",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.MetadataRecord"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/handlers.PendingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Validates a URL, stores a pending metadata record, and either fetches inline or dispatches async collection depending on feature flags.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Create or refresh metadata for a URL",
                "parameters": [
                    {
                        "description": "URL to inventory",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.MetadataRecord"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/db.MetadataRecord"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/db.MetadataRecord"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apperrors.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "db.CookieRecord": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "http_only": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "secure": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "db.MetadataRecord": {
            "type": "object",
            "properties": {
                "cookies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.CookieRecord"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "fetch_duration_ms": {
                    "type": "integer"
                },
                "fetched_at": {
                    "type": "string"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "page_source": {
                    "type": "string"
                },
                "page_source_size_bytes": {
                    "type": "integer"
                },
                "retry_count": {
                    "type": "integer"
                },
                "schema_version": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/db.Status"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "url_hash": {
                    "type": "string"
                }
            }
        },
        "db.Status": {
            "type": "string",
            "enum": [
                "pending",
                "fetching",
                "ready",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusFetching",
                "StatusReady",
                "StatusFailed"
            ]
        },
        "handlers.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "handlers.PendingResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "poll_after_seconds": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handlers.PostRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "handlers.ReadyResponse": {
            "type": "object",
            "properties": {
                "checks": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}